# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Spooled Cloud API
  description: |
    Production-ready webhook queue service API.

    Spooled Cloud provides a reliable, multi-tenant job queue system with:
    - Custom webhook ingestion (receive events from any source)
    - Job queue with retry logic and dead-letter queue
    - Worker coordination with heartbeat and lease management
    - Real-time job status updates via WebSocket/SSE
    - Workflow orchestration with job dependencies
    - Cron-based job scheduling
    - Stripe billing integration for subscription management

    ## Authentication

    All API endpoints (except health checks) require authentication via API key.
    Include your API key in the `Authorization` header:

    ```
    Authorization: Bearer sp_live_your_api_key_here
    ```

    ## Security

    - **Multi-tenant isolation**: All operations are scoped to your organization
    - **API key verification**: Keys verified using bcrypt with caching
    - **Webhook signatures**: HMAC-SHA256 with constant-time comparison
    - **Rate limiting**: Per-key limits with graceful degradation
    - **Input validation**: All inputs sanitized and size-limited
    - **SSRF protection**: Webhook URLs validated in production

    ## Rate Limiting

    Rate limits are applied per API key based on plan tier:
    - Free tier: 5 requests/second (burst: 10)
    - Starter tier: 25 requests/second (burst: 50)
    - Pro tier: 100 requests/second (burst: 200)
...

...

    ## Plan Limits

...

...
  license:
...
servers:
...
paths:
...
  /api/v1/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      description: List jobs with optional filtering
      operationId: listJobs
      parameters:
        - name: queue_name
          in: query
          schema:
            type: string
          description: Filter by queue name
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, scheduled, processing, completed, failed, deadletter, cancelled]
          description: Filter by status
        - name: tag
          in: query
          schema:
            type: string
            maxLength: 64
          description: Filter jobs by a single tag (matches Postgres `tags ? tag` semantics)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Maximum number of results
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset for pagination
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobSummary'
    post:
      tags: [Jobs]
      summary: Create job
      description: |
        Create a new job in the specified queue.

        If an `idempotency_key` is provided and a job with the same key exists,
        the existing job ID is returned (idempotent behavior).
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJobResponse'
        '200':
          description: Existing job returned (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateJobResponse'
  /api/v1/jobs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Job ID
    get:
      tags: [Jobs]
      summary: Get job
      operationId: getJob
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
    delete:
      tags: [Jobs]
      summary: Cancel job
      description: Cancel a pending or scheduled job
      operationId: cancelJob
      responses:
        '204':
          description: Job cancelled
        '404':
          description: Job not found
        '409':
          description: Job cannot be cancelled (not in pending/scheduled state)
...
  /api/v1/queues/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Queue name
    get:
      tags: [Queues]
      summary: Get queue configuration
      operationId: getQueueConfig
      responses:
        '200':
          description: Queue configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueConfig'
        '404':
          description: Queue not found

    delete:
      tags: [Queues]
      summary: Delete queue
      description: |
        Delete a queue and optionally all its jobs.
        Use with caution - this operation is destructive.
      operationId: deleteQueue
      parameters:
        - name: delete_jobs
          in: query
          schema:
            type: boolean
            default: false
          description: Also delete all jobs in the queue
      responses:
        '204':
          description: Queue deleted
        '404':
          description: Queue not found
        '409':
          description: Queue has pending or processing jobs (only when delete_jobs is false)
  /api/v1/queues/{name}/config:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [Queues]
      summary: Update queue configuration
      operationId: updateQueueConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertQueueConfigRequest'
      responses:
        '200':
          description: Queue configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueConfig'
  /api/v1/queues/{name}/stats:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Queues]
      summary: Get queue statistics
      operationId: getQueueStats
      responses:
        '200':
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStats'
...
  /api/v1/events/queues/{name}:
    get:
      tags: [Real-time]
      summary: Queue stats SSE stream
      description: Subscribe to statistics updates for a specific queue
      operationId: sseQueueHandler
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SSE stream for queue stats
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/RealtimeEvent'

  # Schedule (Cron) Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key in Bearer token format
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for platform administration
  schemas:
...
    Job:
      type: object
      required:
        [
          id,
          organization_id,
          queue_name,
          status,
          payload,
          retry_count,
          max_retries,
          timeout_seconds,
          priority,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
        organization_id:
          type: string
        queue_name:
          type: string
        status:
          type: string
          enum: [pending, scheduled, processing, completed, failed, deadletter, cancelled]
        payload:
          type: object
        result:
          type: object
        retry_count:
          type: integer
        max_retries:
          type: integer
        last_error:
          type: string
        created_at:
          type: string
          format: date-time
        scheduled_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        priority:
          type: integer
        tags:
          type: object
        timeout_seconds:
          type: integer
        parent_job_id:
          type: string
        completion_webhook:
          type: string
          format: uri
        assigned_worker_id:
          type: string
        lease_id:
          type: string
        lease_expires_at:
          type: string
          format: date-time
        idempotency_key:
          type: string
        updated_at:
          type: string
          format: date-time
    JobSummary:
      type: object
      required: [id, queue_name, status, priority, attempt, max_retries, job_type, created_at]
      properties:
        id:
          type: string
        queue_name:
          type: string
        status:
          type: string
        priority:
          type: integer
        attempt:
          type: integer
          description: Retry count (serialized as attempt, not retry_count)
        max_retries:
          type: integer
        job_type:
          type: string
          description: From payload.job_type when present; empty string otherwise
        last_error:
          type: string
          nullable: true
          description: Last failure message when the job has failed or dead-lettered
        created_at:
          type: string
          format: date-time
        scheduled_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    CreateJobRequest:
      type: object
      required: [queue_name, payload]
      properties:
        queue_name:
          type: string
          minLength: 1
          maxLength: 100
        payload:
          type: object
        priority:
          type: integer
          minimum: -100
          maximum: 100
          default: 0
        max_retries:
          type: integer
          minimum: 0
          maximum: 100
          default: 3
          description: Omitted → the queue's `max_retries` from `PUT /queues/{name}/config`, else the server default (3)
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 86400
          default: 300
          description: Omitted → the queue's `default_timeout`, else the server default (300)
        scheduled_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        tags:
          type: object
        parent_job_id:
          type: string
        completion_webhook:
          type: string
          format: uri
        completion_webhook_secret:
          type: string
          description: Optional shared secret used to sign completion webhook callbacks.
        idempotency_key:
          type: string
          maxLength: 255
    CreateJobResponse:
      type: object
      required: [id, created]
      properties:
        id:
          type: string
          description: Job ID (new or existing if idempotent)
        created:
          type: boolean
          description: Whether this was a new job or existing
...
security:
  - bearerAuth: []